update mksurfdata in new iac code
This enables the iac component to generate land surface files
that are consistent with the current ELM. A few bugs were also
fixed. The gcam wood harvest is converted to MgC.
Note that when no pft changes are applied there currently can
be differences in land file pft percents between the first and second
years of <0.003. The GLM outputs end up with fraction
differences on the order of 1e-7, and these propagate through
the rest of the code.