Updated passing of pft fraction of grid cell values
Now all pft weights are passed as fraction of actual grid cell. Note
that the coupler variables still are labelled with pct. These
fractions appear to be transferred correctly to land runs. Some
controls have also been updated, particularly so that scalars can
be calculated and written even if they are not applied to gcam.
Also, the iac domain is defined based on exchange with the
atmosphere.

	modified:   cpl/iac_comp_mct.F90
	modified:   cpl/iac_import_export.F90
	modified:   iac/coupling/gcam2glm_mod.F90
	modified:   iac/coupling/gcam_comp_mod.F90
	modified:   iac/coupling/gcam_var_mod.F90
	modified:   iac/coupling/iac_data_mod.F90
	modified:   iac/coupling/iac_init_mod.F90
	modified:   iac/gcam