Fixed harvest normalization bug
The primary and secondary land were not being read in
updateannuallanduse to properly normalize the harvest fraction.
This is mysterious because this code has worked fine for 3 years
in standalong mode. I had to make a local variable to control the loop
because it changed in the loop and didn't allow the proper variables
to be found.