Fixed gcam2glm bugs
The setting of GCAM projected data is now dependent on the five year
GCAM run cycle, otherwise these data get reset with init values.
It is possible that most of the gcam2glm code could be run only
when GCAM runs.
The full available area for ag expansion is now included in the
transition processing, which means that the newly abadoned area is
now available.
The conditions where not enough land is available have been changed
from aborting the code to limiting the expansion.