Increase efficiency and update carbon scaling
This part of the code is mainly about increasing efficiency.
The main processing in gcam2glm_run now runs only in years when
GCAM runs. The interpolation to each year happens each year.
Mksurfdata now fills the pft and harvest data in memory, rather
than it being read from the dynamic landuse timeseries file afterwards.
This required the dyn landuse file name to be a namelist variable.
The pft data are also moved to the previous value array each year.